refactor(nvim): move diagnostics setup to core.lsp
